requirements.txt
fastapi
uvicorn
qgis
pyjwt
httpx<1.0
python-multipart
